Move steps to jbi.actions.steps
